requirements.txt
yt-dlp>=2024.10.7
PySide6>=6.6.0
click>=8.1.7
requests>=2.31.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.0
tabulate>=0.9.0
